requirements.txt
# FastAPI and Server
fastapi==0.109.2
uvicorn[standard]==0.27.1
gunicorn==21.2.0
python-multipart==0.0.9

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
alembic==1.13.1

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
argon2-cffi==23.1.0
cryptography==42.0.2
pyotp==2.9.0
qrcode==7.4.2

# Validation and Serialization
pydantic==2.6.1
pydantic-settings==2.1.0
email-validator==2.1.0.post1

# Redis for Rate Limiting and Sessions
redis==5.0.1
aioredis==2.0.1

# HTTP Client
httpx==0.26.0
aiohttp==3.9.3

# Logging and Monitoring
structlog==24.1.0
python-json-logger==2.0.7
prometheus-client==0.19.0

# Rate Limiting
slowapi==0.1.9

# Security Headers
secure==0.3.0

# UUID Generation
uuid6==2024.1.12

# Environment Variables
python-dotenv==1.0.1

# Testing
pytest==7.0.0
pytest-asyncio==0.23.4
pytest-cov==4.1.0
httpx==0.26.0

# Type Hints
typing-extensions==4.9.0
